backend/requirements.txt
numpy>=1.25.2,<2.0
Flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
pymongo>=4.6.1,<5.0.0
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.1.2,<5.0.0
opencv-python>=4.9.0.80,<5.0.0
Pillow>=10.2.0,<11.0.0
transformers>=4.36.2,<5.0.0
torch>=2.2.0,<2.3.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
